# static/swagger.spec.json
{
  "info": {
    "title": "gearlaunch2-backend-rest-demo",
    "version": "1.0.0",
    "description": "GearLaunch v2 Rest Demo Backend"
  },
  "host": "localhost:8000",
  "schemes": [
    "http",
    "https"
  ],
  "basePath": "/",
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "swagger": "2.0",
  "paths": {
...
    "/api/products/v1/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Find all Products",
        "description": "This API returns a list of all `Products` based on common `select` request.\n",
        "operationId": "findAllProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get a list of all `Products`, you can submit this `select`:\n```\n{\n  \"name\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Products` with status `in_stock` that were created by `Merchants` from Africa, you can submit this `select`:\n```\n{\n  \"name\": \"\",\n  \"status\": \"\",\n  \"$where\": {\n    \"status\": {\n      \"$eq\": \"in_stock\"\n    }\n  },\n  \"merchant\": {\n    \"merchant_name\": \"\",\n    \"country\": {\n      \"name\": \"\",\n      \"continent_name\": \"\",\n      \"$where\": {\n        \"continent_name\": {\n          \"$eq\": \"Africa\"\n        }\n      }\n    }\n  }\n}\n```\n",
            "default": {
              "name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of `Products` that matched the request parameters.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DatabaseProduct"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
...
    "/api/accounts/v1/users": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Create a User",
        "description": "Take the incoming POST Body and try to create a `User` entity out of it.\nThis API will return the newly created `User` on success.\n",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "The payload data for the `User` we want to create.\n",
            "schema": {
              "type": "object",
              "required": [
                "full_name",
                "country_code"
              ],
              "properties": {
                "full_name": {
                  "type": "string",
                  "minimum": 2,
                  "example": "John Smith"
                },
                "country_code": {
                  "type": "integer",
                  "minimum": 1,
                  "example": 21
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The newly created User",
            "schema": {
              "$ref": "#/definitions/DatabaseUser"
            }
          },
          "400": {
            "$ref": "#/responses/InvalidRequest"
          },
          "404": {
            "description": "Unable to find Country with the requested code.",
            "schema": {
              "$ref": "#/definitions/ResponseError"
            },
            "examples": {
              "application/json": {
                "errorId": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7",
                "errorCode": "COUNTRY_NOT_FOUND",
                "errorMessage": "Unable to find country"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      },
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Find all Users",
        "description": "This API returns a list of all `Users` based on common `select` request.\n",
        "operationId": "findAllUsers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get a list of all `Users` with their full name, you can submit this `select`:\n```\n{\n  \"full_name\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Users` within the continent of `Africa`, you can submit this `select`:\n```\n{\n  \"full_name\": \"\",\n  \"country\": {\n    \"name\": \"\",\n    \"continent_name\": \"\",\n    \"$where\": {\n      \"continent_name\": {\n        \"$eq\": \"Africa\"\n      }\n    }\n  }\n}\n```\n",
            "default": {
              "full_name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of `Users` that matched the request parameters.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DatabaseUser"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
    "/api/accounts/v1/users/{id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Find one User",
        "description": "This API returns a single one `User` based on common `select` request\nand the `id` URL param\n",
        "operationId": "findOneUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/PathUniqueIdentifier"
          },
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get the `User` full name, you can submit this `select`:\n```\n{\n  \"full_name\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Products` a specific `User` bought around Christmas, you can submit this `select`:\n```\n{\n  \"full_name\": \"\",\n  \"orders\": {\n    \"created_at\": \"\",\n    \"$where\": {\n      \"created_at\": {\n        \"$and\": [\n          {\"$gt\": \"2019-12-10T00:00:00.000Z\"},\n          {\"$lt\": \"2019-12-27T00:00:00.000Z\"}\n        ]\n      }\n    },\n    \"order_items\": {\n      \"quantity\": \"\",\n      \"product\": {\n        \"name\": \"\"\n      }\n    }\n  }\n}\n```\n",
            "default": {
              "full_name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single one `Users` that matched the request parameters.\n",
            "schema": {
              "$ref": "#/definitions/DatabaseUser"
            }
          },
          "404": {
            "description": "Unable to find User with the requested params.",
            "schema": {
              "$ref": "#/definitions/ResponseError"
            },
            "examples": {
              "application/json": {
                "errorId": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7",
                "errorCode": "USER_NOT_FOUND",
                "errorMessage": "Unable to find user"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    }
  },
  "definitions": {
    "ResponseError": {
      "type": "object",
      "description": "Schema for `4XX` server error response body.\nResponse includes a UUIDv4, error code and message body.\n",
      "properties": {
        "errorId": {
          "type": "string",
          "format": "uuid",
          "example": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7"
        },
        "errorCode": {
          "type": "string",
          "example": "GENERIC_SERVER_ERROR"
        },
        "errorMessage": {
          "type": "string",
          "example": "Generic server error"
        }
      },
      "required": [
        "errorId",
        "errorCode",
        "errorMessage"
      ]
    },
...
    "DatabaseProduct": {
      "description": "`Products` are entities that `merchants` preset as ready to be sold.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "id",
            "name",
            "merchant_id",
            "price",
            "status",
            "created_at"
          ],
          "properties": {
            "id": {
              "type": "number",
              "description": "Unique merchant ID",
              "example": 34
            },
            "name": {
              "type": "string",
              "description": "Product UI representation",
              "example": "Jeans"
            },
            "merchant_id": {
              "type": "number",
              "description": "Association to the `merchants` entity. Shows which merchants sells this product.",
              "example": 14
            },
            "price": {
              "type": "number",
              "description": "How much does this product costs in USD",
              "example": 100.25
            },
            "status": {
              "type": "string",
              "description": "Indicates if this product is available for selling. Possible values are out_of_stock, in_stock, running_low. Default value is out_of_stock.",
              "example": "in_stock",
              "default": "out_of_stock"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that product was created",
              "example": "2020-04-30T00:00:00.000Z"
            }
          }
        }
      ]
    },
    "DatabaseUser": {
      "description": "`Users` are the administrators that oversee trade. `Users` can also sell `products`\nif they have a relation to a `merchant` entity.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "id",
            "full_name",
            "country_code",
            "created_at"
          ],
          "properties": {
            "id": {
              "type": "number",
              "description": "Unique user ID",
              "example": 34
            },
            "full_name": {
              "type": "string",
              "description": "User names",
              "example": "John Smith"
            },
            "country_code": {
              "type": "number",
              "description": "Association to the country entity. Shows user country address.",
              "example": 22
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that user was created",
              "example": "2020-04-30T00:00:00.000Z"
            }
          }
        }
      ]
    }
  },
  "responses": {
    "InvalidRequest": {
      "description": "Invalid parameters sent to the server.\n",
      "schema": {
        "$ref": "#/definitions/ResponseError"
      },
      "examples": {
        "application/json": {
          "errorId": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7",
          "errorCode": "REQUEST_VALIDATION_ERROR",
          "errorMessage": "Missing request \"${type}\""
        }
      }
    },
    "AppNotFound": {
      "description": "Some unexpected issue occurred and crashed the app.\n",
      "schema": {
        "$ref": "#/definitions/ResponseError"
      },
      "examples": {
        "application/json": {
          "errorId": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7",
          "errorCode": "INTERNAL_SERVER_ERROR",
          "errorMessage": "Internal server error"
        }
      }
    }
  },
  "parameters": {
...
    "PathUniqueIdentifier": {
      "in": "path",
      "name": "id",
      "type": "number",
      "format": "integer",
      "required": true,
      "description": "The `id` property of the entity we want to retrieve.\n"
    },
...
  },
...
}